fix: split skill hub client api
