+++ apps/client/src/api/knowledge.ts
@@ -62,51 +62,6 @@ export interface SkillDetail extends SkillSummary {
   body: string
 }
 
-export interface SkillHubRegistrySummary {
-  id: string
-  type: 'claude-code' | 'skills-sh'
-  enabled: boolean
-  searchable: boolean
-  source: string
-  pluginCount?: number
-  error?: string
-}
-
-export interface SkillHubItem {
-  id: string
-  registry: string
-  name: string
-  description?: string
-  version?: string
-  skills: string[]
-  commands: string[]
-  agents: string[]
-  mcpServers: string[]
-  hasHooks: boolean
-  installed: boolean
-  installScope?: string
-  installedAt?: string
-  installRef?: string
-  source?: string
-  detailUrl?: string
-  installs?: number
-}
-
-export interface SkillHubSearchResult {
-  hasMore?: boolean
-  registries: SkillHubRegistrySummary[]
-  items: SkillHubItem[]
-}
-
-export interface SkillHubInstallResult {
-  registry: string
-  plugin: string
-  name: string
-  scope?: string
-  installedAt: string
-  installDir: string
-}
-
 export async function listSpecs(): Promise<{ specs: SpecSummary[] }> {
   return fetchApiJson<{ specs: SpecSummary[] }>('/api/ai/specs')
 }
@@ -172,35 +127,4 @@ export async function getSkillDetail(path: string): Promise<{ skill: SkillDetail
   const url = createApiUrl('/api/ai/skills/detail')
   url.searchParams.set('path', path)
   return fetchApiJson<{ skill: SkillDetail }>(url)
-}
-
-export async function searchSkillHub(params: {
-  limit?: number
-  registry?: string
-  query?: string
-} = {}): Promise<SkillHubSearchResult> {
-  const url = createApiUrl('/api/skill-hub/search')
-  if (params.limit != null) {
-    url.searchParams.set('limit', String(params.limit))
-  }
-  if (params.registry != null && params.registry !== '') {
-    url.searchParams.set('registry', params.registry)
-  }
-  if (params.query != null && params.query !== '') {
-    url.searchParams.set('q', params.query)
-  }
-  return fetchApiJson<SkillHubSearchResult>(url)
-}
-
-export async function installSkillHubItem(params: {
-  registry: string
-  plugin: string
-  force?: boolean
-  scope?: string
-}): Promise<SkillHubInstallResult> {
-  return fetchApiJson<SkillHubInstallResult>('/api/skill-hub/install', {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify(params)
-  })
 }
